Don't advertise a valid power-supply via SOURCE_CAP
On a M2 MBA, disconnecting the power source results in a constant
PD renegotiation, as we advertise a power supply but never provide
any current (we advertise 5V, 1.5A... good luck with that).

Change the advertised source capability to be a non-battery variable
power supply that has nothing to offer (0V, 0mA). This ensures that
the DUT will not try to source anything from it, and thus doesn't
renegociate. See 6.4.1.2.4 "Variable Supply (non-Battery) Power Data
Object" in the USB-PD 2.0 spec.

This certainly seem to work with my M2 MBA, but it should be obvious
to anyone that I have no idea what I'm doing here!

Signed-off-by: Marc Zyngier <maz@kernel.org>