Clear CC pull-ups before starting measurements
Despite the comment, the pull-ups do not get cleared when starting
the measurement. Which means that a measurement can influence another
by leaving the pull-up from a previous measurement set.

This manifests itself by both CC lines always having the same values,
which is a bit odd.

Clearing the PUs fixes the measurement, and we now correctly report
the connector orientation (on a board hacked to use its CC2 line,
obviously).

Signed-off-by: Marc Zyngier <maz@kernel.org>