Add support for USB2.0 routing
CSv3 has a switch allowing the routing of the USB2.0 pins to
either the external USB (as for all th eprevious versions)
or the serial lines (to support dumb cables).

The switch is controlled by a GPIO, which needs to be set to
the right level when claming the serial port.

Signed-off-by: Marc Zyngier <maz@kernel.org>