Increase delay when probing CC lines
It appears that the FUSB302 A_revB is getting confused when reading
the CC lines back to back, and reports a whole lot of crap.

Bumping the delay from 200ms to 500ms seems to help, so here it is.

Also display the device ID in a format that matches the spec...

Signed-off-by: Marc Zyngier <maz@kernel.org>