Don't measure CC lines twice
Doing a double measurement is a waste of time, and can result in
lost messages if the machine boots faster than the FUSB302 replies
(it is the case on my M2 Pro, where I sometimes lose the initial
m1n1 boot).

Also makes the sleeping variable depending on the FUSB302 version,
as the version B seems to be significantly faster.

Signed-off-by: Marc Zyngier <maz@kernel.org>